+++ Sources/SentryObjC/Public/SentryObjCDataCollectionDatabaseCollectionOptions.h
@@ -1,4 +1,5 @@
-#import <Foundation/Foundation.h>
+#if SDK_V10
+#    import <Foundation/Foundation.h>
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -13,3 +14,4 @@ NS_ASSUME_NONNULL_BEGIN
 @end
 
 NS_ASSUME_NONNULL_END
+#endif // SDK_V10