+++ Sources/SentryObjC/Public/SentryObjCDataCollectionOptions.h
@@ -1,9 +1,10 @@
-#import <Foundation/Foundation.h>
-#if !__has_include(<SentryObjC/SentryObjCDefines.h>)
-#    import "SentryObjCDataCollectionHttpBodyType.h"
-#else
-#    import <SentryObjC/SentryObjCDataCollectionHttpBodyType.h>
-#endif
+#if SDK_V10
+#    import <Foundation/Foundation.h>
+#    if !__has_include(<SentryObjC/SentryObjCDefines.h>)
+#        import "SentryObjCDataCollectionHttpBodyType.h"
+#    else
+#        import <SentryObjC/SentryObjCDataCollectionHttpBodyType.h>
+#    endif
 
 @class SentryObjCDataCollectionKeyValueCollectionBehavior;
 @class SentryObjCDataCollectionHttpHeaderCollectionOptions;
@@ -57,3 +58,4 @@ NS_ASSUME_NONNULL_BEGIN
 @end
 
 NS_ASSUME_NONNULL_END
+#endif // SDK_V10