+++ Sources/SentryObjCCompat/SentryObjCDataCollectionKeyValueCollectionBehavior.swift
@@ -1,3 +1,4 @@
+#if SDK_V10
 // swiftlint:disable missing_docs
 #if SWIFT_PACKAGE
 internal import SentrySwift
@@ -54,3 +55,4 @@ public final class SentryObjCDataCollectionKeyValueCollectionBehavior: NSObject 
 }
 
 // swiftlint:enable missing_docs
+#endif // SDK_V10