+++ Sources/SentryObjCCompat/SentryObjCDataCollectionKeyValueCollectionMode.swift
@@ -1,3 +1,4 @@
+#if SDK_V10
 // swiftlint:disable missing_docs
 
 @objc public enum SentryObjCDataCollectionKeyValueCollectionMode: Int {
@@ -7,3 +8,4 @@
 }
 
 // swiftlint:enable missing_docs
+#endif // SDK_V10