+++ Tests/SentryTests/DataCollection/SentryDatabaseCollectionOptionsTests.swift
@@ -3,25 +3,36 @@ import XCTest
 
 class SentryDatabaseCollectionOptionsTests: XCTestCase {
 
-    func testInit_withoutArguments_shouldDefaultToTrue() {
+    func testInit_withoutArguments_shouldDefaultToTrue() throws {
+        #if !SDK_V10
+        throw XCTSkip("Test skipped for SDK_V10")
+        #else
         // -- Act --
         let options = SentryDataCollection.DatabaseCollectionOptions()
 
         // -- Assert --
         XCTAssertTrue(options.queryParams)
+        #endif
     }
 
-    func testInit_withArguments_shouldSetProperties() {
+    func testInit_withArguments_shouldSetProperties() throws {
+        #if !SDK_V10
+        throw XCTSkip("Test skipped for SDK_V10")
+        #else
         // -- Act --
         let options = SentryDataCollection.DatabaseCollectionOptions(queryParams: false)
 
         // -- Assert --
         XCTAssertFalse(options.queryParams)
+        #endif
     }
 
     // MARK: - Dictionary Init
 
-    func testInitWithDictionary_whenQueryParamsIsPresent_shouldSetQueryParams() {
+    func testInitWithDictionary_whenQueryParamsIsPresent_shouldSetQueryParams() throws {
+        #if !SDK_V10
+        throw XCTSkip("Test skipped for SDK_V10")
+        #else
         // -- Arrange --
         let dictionary: [String: Any] = ["queryParams": false]
 
@@ -30,9 +41,13 @@ class SentryDatabaseCollectionOptionsTests: XCTestCase {
 
         // -- Assert --
         XCTAssertFalse(options.queryParams)
+        #endif
     }
 
-    func testInitWithDictionary_whenQueryParamsIsMissing_shouldUseDefault() {
+    func testInitWithDictionary_whenQueryParamsIsMissing_shouldUseDefault() throws {
+        #if !SDK_V10
+        throw XCTSkip("Test skipped for SDK_V10")
+        #else
         // -- Arrange --
         let dictionary: [String: Any] = [:]
 
@@ -41,9 +56,13 @@ class SentryDatabaseCollectionOptionsTests: XCTestCase {
 
         // -- Assert --
         XCTAssertEqual(options, SentryDataCollection.DatabaseCollectionOptions())
+        #endif
     }
 
-    func testInitWithDictionary_whenQueryParamsIsNSNull_shouldUseDefault() {
+    func testInitWithDictionary_whenQueryParamsIsNSNull_shouldUseDefault() throws {
+        #if !SDK_V10
+        throw XCTSkip("Test skipped for SDK_V10")
+        #else
         // -- Arrange --
         let dictionary: [String: Any] = ["queryParams": NSNull()]
 
@@ -52,9 +71,13 @@ class SentryDatabaseCollectionOptionsTests: XCTestCase {
 
         // -- Assert --
         XCTAssertEqual(options, SentryDataCollection.DatabaseCollectionOptions())
+        #endif
     }
 
-    func testInitWithDictionary_whenQueryParamsHasWrongType_shouldUseDefault() {
+    func testInitWithDictionary_whenQueryParamsHasWrongType_shouldUseDefault() throws {
+        #if !SDK_V10
+        throw XCTSkip("Test skipped for SDK_V10")
+        #else
         // -- Arrange --
         let dictionary: [String: Any] = ["queryParams": "false"]
 
@@ -63,5 +86,6 @@ class SentryDatabaseCollectionOptionsTests: XCTestCase {
 
         // -- Assert --
         XCTAssertEqual(options, SentryDataCollection.DatabaseCollectionOptions())
+        #endif
     }
 }